Fix inf loop in ListBundleFormat#createNewField().
I not a fan of the ListBundleFormat#createNewField() method.
It is kind of silly it allows the user to create a field that
is guaranteed to be a new field. There was a pathological case
where this method could encounter an infinite loop. This commit
fixes the infinite loop by incrementing suffix to the next integer.
Technically it is still possible to generate an infinite loop by
creating 2^32 new bundle fields in a bundle but you have other
problems if you are doing this.