derived typed field should return null for null
I think the idea previously was to let jackson use whatever it
defined as null for that type was, but calling "asNative()" would
stop that from happening. We could special case null to null and
then pass it in to jackson anyway, but this seems easier for users
to reason about for now.