add asMap, asValueMap, and asValueArray to Bundle
these are all given default (albeit possibly rather inefficient)
implementations to protect downstream implementations from any
breaking changes. asMap returns a view of the bundle as a map of
BundleFields to ValueObjects, and the other two return it as value
objects of the corresponding type. All such views support all
mutation operations available to those types, and all changes in
either the view or the bundle are reflected in the other.